+++ src/main/java/site/moamoa/backend/repository/CategoryRepository.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.web.repository;
+package site.moamoa.backend.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import site.moamoa.backend.domain.Category;
