+++ src/main/java/site/moamoa/backend/repository/PostRepository.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.web.repository;
+package site.moamoa.backend.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
