+++ src/main/java/site/moamoa/backend/service/MemberPostCommandService.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.web.service;
+package site.moamoa.backend.service;
 
 import site.moamoa.backend.domain.mapping.MemberPost;
