+++ src/main/java/site/moamoa/backend/service/MemberPostQueryService.java
@@ -1,4 +1,4 @@
-package site.moamoa.backend.web.service;
+package site.moamoa.backend.service;
 
 public interface MemberPostQueryService {
